Assistant checkpoint: Created VIP Program page with tier benefits and routing
Assistant generated file changes:
- client/src/pages/VipProgram.tsx: Create VIP Program page with tier benefits
- client/src/App.tsx: Add VIP Program route
Add VIP Program route

---

User prompt:

We need to create the VIP PROGRAM page, that shows info on the vip program and tiers, EARN REWARDS WITH GOATED VIP PROGRAM
Goated's VIP program is tailored to accommodate all types of players, with a focus on maximizing the cumulative bonuses you receive for every dollar wagered. i will upload a file of a table we can display on this page as well as some info on the program. also, ive added a folder to the client>public>images directory called (Goated Emblems) this has all the icons of the current vip tiers including; bronze, silver, gold, platinum, pearl, saphire emerald and diamond, we should showcase these on this page too. link this page to our VIP PROGRAM tab in the top bar, and also to the VIP REWARDS tooltip box with the explore vip program navigation link.